Extract useScreenAnimatePresence hook, tidy up
